Add /user/:user_id API
